bpf, docs: Move the introduction to eBPF into a separate document
Split the introductory that explain eBPF vs classic BPF and how it maps
to hardware from the instruction set specification into a standalone
document.  Because this introduction was the only place explaining the
register set and calling conventins keep those in the main instruction
set document after a small refactoring.

Signed-off-by: Christoph Hellwig <hch@lst.de>