selftests/bpf: CGROUP_SKB test for skb_store_bytes()
Test that the source port of an egress packet is changed
via skb_store_bytes() for a BPF_PROG_TYPE_CGROUP_SKB prog.

Test creates a client and server and checks that the packet
received on the server has the updated source IP and Port
that the bpf program modifies.

Signed-off-by: Tyler Wear <quic_twear@quicinc.com>