ice: make Tx threshold dependent on ring length
XDP_TX workloads use a concept of Tx threshold that indicates the
interval of setting RS bit on descriptors which in turn tells the HW to
generate an interrupt to signal the completion of Tx on HW side. It is
currently based on a constant value of 32 which might not work out well
for various sizes of ring combined with for example batch size that can
be set via SO_BUSY_POLL_BUDGET.

Internal tests based on AF_XDP showed that most convenient setup of
mentioned threshold is when it is equal to quarter of a ring length.

Introduce @tx_thresh field in ice_tx_ring struct that will store the
value of threshold and use it in favor of ICE_TX_THRESH.

Signed-off-by: Maciej Fijalkowski <maciej.fijalkowski@intel.com>